feat: add permanent links to conversations - make messages more readable on mobile - refactor AiAssistant page to be routable
