+++ plugins/ai-assistant/src/components/ConversationPage/index.tsx
@@ -1,0 +1,1 @@
+export * from './ConversationPage';
