+++ plugins/ai-assistant/src/routes.ts
@@ -1,5 +1,17 @@
-import { createRouteRef } from '@backstage/core-plugin-api';
+import { createRouteRef, createSubRouteRef } from '@backstage/core-plugin-api';
 
 export const rootRouteRef = createRouteRef({
   id: 'ai-assistant',
+});
+
+export const newConversationRouteRef = createSubRouteRef({
+  id: 'ai-assistant.new-conversation',
+  parent: rootRouteRef,
+  path: '/conversation',
+});
+
+export const conversationRouteRef = createSubRouteRef({
+  id: 'ai-assistant.conversation',
+  parent: rootRouteRef,
+  path: '/conversation/:id',
 });
